Remove CONLL_TO_LABEL_MAP; move encode to roberta_util
