+++ .gitignore
@@ -3,3 +3,4 @@ __pycache__
 *.log
 wifineparse
 fewnerdparse
+glove.6B
