+++ roberta_util.py
@@ -1,8 +1,5 @@
 import logging
-import random
-#from typing import Collection, Optional, Union
 
-import torch
 from transformers import RobertaForTokenClassification, RobertaTokenizer
 
 
@@ -27,3 +24,21 @@ def make_model(num_labels, id2label, label2id):
 def make_tokenizer():
     tokenizer = RobertaTokenizer.from_pretrained(ROBERTA_VERSION)
     return tokenizer
+
+
+def encode(dataset, tokenizer):
+    def add_encodings(example):
+
+        # get the encodings of the tokens. The tokens are already split, that is why we must add is_split_into_words=True
+        encodings = tokenizer(example['tokens'], truncation=True,
+                              padding='max_length', is_split_into_words=True)
+        # extend the ner_tags so that it matches the max_length of the input_ids
+        labels = example['ner_tags'] + [0] * \
+                 (tokenizer.model_max_length - len(example['ner_tags']))
+        # return the encodings and the extended ner_tags
+        return {**encodings, 'labels': labels}
+
+    dataset = dataset.map(add_encodings)
+    dataset.set_format(type='torch', columns=[
+        'input_ids', 'attention_mask', 'labels'])
+    return dataset
